+++ src/main/java/com/clinicadental/clinica/repository/IAddressRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAddressRepository extends JpaRepository<Address, Integer> {
+public interface IAddressRepository extends JpaRepository<Address, Long> {
 
 }
