+++ src/main/java/com/clinicadental/clinica/repository/IDentistRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IDentistRepository extends JpaRepository<Dentist, Integer> {
+public interface IDentistRepository extends JpaRepository<Dentist, Long> {
 
 }
